+++ packages/security/src/errors/index.ts
@@ -1,2 +1,3 @@
+export * from './password-complexity-error'
 export * from './unauthenticated-error'
 export * from './unauthorized-error'
