+++ packages/security/src/models/index.ts
@@ -1,3 +1,5 @@
-export * from './password-credential'
-export * from './password-complexity-rule'
+export * from './password-check-result'
 export * from './password-complexity-result'
+export * from './password-complexity-rule'
+export * from './password-credential'
+export * from './password-reset-token'
