+++ packages/security/src/password-complexity-rules/contains-lowercase.spec.ts
@@ -1,4 +1,4 @@
-import { PasswordComplexityFailedResult } from '..'
+import { PasswordComplexityFailedResult } from '../models'
 import { createContainsLowercasePolicy } from './contains-lowercase'
 
 describe('Contains-lowercase', () => {
