+++ packages/security/src/password-complexity-rules/max-length.ts
@@ -1,4 +1,4 @@
-import type { PasswordComplexityRule } from '../models/password-complexity-rule'
+import type { PasswordComplexityRule } from '../models'
 
 /**
  * @param maxLength The maximum password length
